cli: Move Constants and Extensions to the utils package
These are not commands.
